renamed the clusterStatusMonitoring job to clusterMonitroiing and moved it inside existing cluster jobs directory for better management
