+++ Tombolo/server/jobSchedularMethods/clusterJobs.js
@@ -5,16 +5,46 @@ const logger = require('../config/logger');
 const models = require('../models');
 const {
   clusterReachabilityMonitoringInterval,
-  // eslint-disable-next-line no-unused-vars
+  cluster_monitoring_interval,
   clusterContainerizationCheckInterval,
 } = require('../config/monitorings.js');
 
 // Constants
+const CLUSTER_MONITORING_FILE_NAME = 'clusterMonitoring.js';
 const CLUSTER_TIMEZONE_OFFSET = 'clustertimezoneoffset.js';
 const CLUSTER_USAGE_HISTORY_TRACKER = 'submitClusterUsageTracker.js';
 const MONITOR_CLUSTER_REACHABILITY_FILE_NAME = 'monitorClusterReachability.js';
 const CHECK_CLUSTER_CONTAINERIZATION_FILE_NAME =
   'checkIfClusterIsContainerized.js';
+
+// Cluster status monitoring bree job
+async function startClusterStatusMonitoring() {
+  try {
+    let jobName = 'cluster-status-monitoring' + new Date().getTime();
+    this.bree.add({
+      name: jobName,
+      interval: '10s', // For development
+      // interval: `${cluster_monitoring_interval}m`,
+      path: path.join(
+        __dirname,
+        '..',
+        'jobs',
+        'cluster',
+        CLUSTER_MONITORING_FILE_NAME
+      ),
+      worker: {
+        workerData: {
+          jobName: jobName,
+          WORKER_CREATED_AT: Date.now(),
+        },
+      },
+    });
+    this.bree.start(jobName);
+    logger.info('Cluster  monitoring job initialized ...');
+  } catch (err) {
+    logger.error(err.message);
+  }
+}
 
 async function scheduleClusterTimezoneOffset() {
   logger.info('Cluster timezone offset checker job initialized ...');
@@ -125,4 +155,5 @@ module.exports = {
   createClusterUsageHistoryJob,
   checkClusterReachability,
   checkClusterContainerization,
+  startClusterStatusMonitoring,
 };
